Correct Pioneer protocol based on real data. * Remove inter-segment protocol message. * Add better address/command decoding to encode/decode routines. * Unit tests based on real example messages.

+++ src/IRsend.h
@@ -302,7 +302,7 @@ void send(uint16_t type, uint64_t data, uint16_t nbits);
 #if SEND_PIONEER
   void sendPioneer(const uint64_t data, const uint16_t nbits = kPioneerBits,
                    const uint16_t repeat = kNoRepeat);
-  uint64_t encodePioneer(uint32_t address, uint32_t command);
+  uint64_t encodePioneer(uint16_t address, uint16_t command);
 #endif
 
  protected:
